+++ bespokeasm/__init__.py
@@ -1,4 +1,4 @@
-BESPOKEASM_VERSION_STR = "0.2.5"
+BESPOKEASM_VERSION_STR = "0.3.0"
 
 # if a cconfig file requires a certain bespoke ASM version, it should be at least this version.
-BESPOKEASM_MIN_REQUIRED_STR = "0.1.4"
+BESPOKEASM_MIN_REQUIRED_STR = "0.3.0"