+++ bespokeasm/assembler/model/instruction_parser.py
@@ -3,8 +3,8 @@ import sys
 from bespokeasm.assembler.model.instruction_parser_base import InstructioParserBase
 from bespokeasm.assembler.model import AssemblerModel
 from bespokeasm.assembler.line_identifier import LineIdentifier
-from bespokeasm.assembler.byte_code.assembled import AssembledInstruction
-from bespokeasm.assembler.byte_code.generator import BytecodeGenerator
+from bespokeasm.assembler.bytecode.assembled import AssembledInstruction
+from bespokeasm.assembler.bytecode.generator import BytecodeGenerator
 from bespokeasm.assembler.memory_zone.manager import MemoryZoneManager