+++ bespokeasm/assembler/model/operand/types/numeric_bytecode.py
@@ -1,7 +1,7 @@
 import sys
 
 from bespokeasm.assembler.line_identifier import LineIdentifier
-from bespokeasm.assembler.byte_code.parts import ExpressionByteCodePartWithValidation
+from bespokeasm.assembler.bytecode.parts import ExpressionByteCodePartWithValidation
 from bespokeasm.assembler.model.operand import Operand, OperandType, ParsedOperand
 from bespokeasm.assembler.memory_zone.manager import MemoryZoneManager
 from bespokeasm.expression import EXPRESSION_PARTS_PATTERN