+++ bespokeasm/assembler/model/operand/types/numeric_enumeration.py
@@ -1,7 +1,7 @@
 import re
 
 from bespokeasm.assembler.line_identifier import LineIdentifier
-from bespokeasm.assembler.byte_code.parts import ExpressionEnumerationByteCodePart
+from bespokeasm.assembler.bytecode.parts import ExpressionEnumerationByteCodePart
 from bespokeasm.assembler.model.operand import OperandType, ParsedOperand
 from bespokeasm.assembler.model.operand.types.numeric_expression import NumericExpressionOperand
 from bespokeasm.assembler.memory_zone.manager import MemoryZoneManager